requirements.txt
requests>=2.31.0
PyYAML>=6.0
usps-api>=0.1.0
Flask>=2.3.0 